fix: types error
