+++ examples/basic/package.json
@@ -38,7 +38,7 @@
     "jest": "^29.6.2",
     "ts-jest": "^29.1.1",
     "typescript": "latest",
-    "vite": "^5.0.13"
+    "vite": "catalog:"
   },
   "engines": {
     "node": ">=18"
