+++ examples/vue/package.json
@@ -39,7 +39,7 @@
     "jest": "^29.6.2",
     "ts-jest": "^29.1.1",
     "typescript": "latest",
-    "vite": "^5.0.13"
+    "vite": "catalog:"
   },
   "engines": {
     "node": ">=18"
