+++ examples/web-router/entry.server.ts
@@ -1,24 +1,18 @@
-import { mergeMeta } from '@web-widget/react';
-import type { Manifest, StartOptions } from '@web-widget/web-router';
+import { mergeMeta } from '@web-widget/helpers';
+import { meta, manifest } from '@web-widget/helpers/placeholder';
 import WebRouter from '@web-widget/web-router';
 
-export default (manifest: Manifest, options: StartOptions) => {
-  return WebRouter.fromManifest(manifest, {
-    ...options,
-    defaultMeta: mergeMeta(
+export default WebRouter.fromManifest(manifest, {
+  defaultMeta: mergeMeta(meta, {
+    lang: 'en',
+    meta: [
       {
-        lang: 'en',
-        meta: [
-          {
-            charset: 'utf-8',
-          },
-          {
-            name: 'viewport',
-            content: 'width=device-width, initial-scale=1.0',
-          },
-        ],
+        charset: 'utf-8',
       },
-      options.defaultMeta || {}
-    ),
-  });
-};
+      {
+        name: 'viewport',
+        content: 'width=device-width, initial-scale=1.0',
+      },
+    ],
+  }),
+});
