+++ examples/web-router/package.json
@@ -19,31 +19,31 @@
   "dependencies": {
     "@examples/web-router-vue2": "workspace:*",
     "@examples/web-router-vue3": "workspace:*",
+    "@web-widget/helpers": "workspace:*",
     "@web-widget/html": "workspace:*",
     "@web-widget/node": "workspace:*",
     "@web-widget/react": "workspace:*",
-    "@web-widget/helpers": "workspace:*",
     "@web-widget/web-router": "workspace:*",
     "@web-widget/web-widget": "workspace:*",
-    "koa": "^2.14.2",
+    "koa": "^2.15.3",
     "koa-connect": "^2.1.0",
     "koa-send": "^5.0.1",
     "react": "rc",
     "react-dom": "rc"
   },
   "devDependencies": {
-    "@web-widget/vite-plugin": "workspace:*",
-    "@types/koa": "^2.13.6",
-    "@types/koa-send": "^4.1.3",
+    "@jest/globals": "^29.7.0",
+    "@types/jest": "^29.5.12",
+    "@types/koa": "^2.15.0",
+    "@types/koa-send": "^4.1.6",
     "@types/react": "npm:types-react@rc",
     "@types/react-dom": "npm:types-react-dom@rc",
-    "@vitejs/plugin-react": "^4.1.1",
+    "@vitejs/plugin-react": "^4.3.1",
+    "@web-widget/vite-plugin": "workspace:*",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.2.4",
     "typescript": "latest",
-    "vite": "^5.0.13",
-    "@jest/globals": "^29.6.2",
-    "@types/jest": "^29.5.3",
-    "jest": "^29.6.2",
-    "ts-jest": "^29.1.1"
+    "vite": "catalog:"
   },
   "engines": {
     "node": ">=18"
