+++ examples/web-widget/package.json
@@ -27,7 +27,7 @@
     "@types/react-dom": "^18.2.18",
     "@vitejs/plugin-react": "^4.2.1",
     "typescript": "latest",
-    "vite": "^5.0.13"
+    "vite": "catalog:"
   },
   "engines": {
     "node": ">=18"
