+++ examples/web-widget/packages/vue3/package.json
@@ -15,6 +15,7 @@
     "vue": "^3.4.8"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^5.0.0"
+    "@vitejs/plugin-vue": "^5.0.0",
+    "vite": "catalog:"
   }
 }
