+++ packages/vite-plugin/package.json
@@ -46,12 +46,12 @@
     "rollup": "^4.9.3",
     "tsup": "^8.0.1",
     "typescript": "latest",
-    "vite": "^5.0.13",
+    "vite": "catalog:",
     "jest": "^29.6.2",
     "ts-jest": "^29.1.1"
   },
   "peerDependencies": {
-    "vite": "^5.0.13"
+    "vite": "^5.4.0"
   },
   "engines": {
     "node": "^14.18 || >= 16.0.0"
